+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [4.1.1] - 2026-02-17
+
+- Update DataShare to function like Dalamuds
+- Update MockCharacter to include CustomizeData
+
 ## [4.1.0] - 2026-02-15
 
 - MockContainer's serviceReplacements dictionary now expects a <InterfaceType, ImplementingType>. The interface should be the dalamud service you want to provide. The implementing type should implement the interface plus IMockService.